refactoring and renaming
refactor index into shared and platform files, and explicitly suffix node files
