+++ common/lib/common-utils/src/index.ts
@@ -11,7 +11,6 @@ export * from "./eventForwarder";
 export * from "./hashFile";
 export * from "./heap";
 export * from "./logger";
-export * from "./performanceNow";
 export * from "./promiseCache";
 export * from "./promises";
 export * from "./rangeTracker";
