+++ common/lib/common-utils/src/indexNode.ts
@@ -4,4 +4,4 @@
  */
 
 export * from "./index";
-export * from "./performanceNowBrowser";
+export * from "./performanceNowNode";
